# libs/api-client/model/loginDto.ts
/**
 * Team Board API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LoginDto { 
    email: string;
    password: string;
}